Breaking: Extract compat information to new package
Fix #3035
Close #3106
