+++ packages/utils-compat-data/.gitignore
@@ -1,0 +1,1 @@
+src/browser-compat-data.ts
