+++ packages/utils/package.json
@@ -23,13 +23,10 @@
     "is-wsl": "^2.1.1",
     "jsonc-parser": "^2.1.1",
     "lodash": "^4.17.15",
-    "mdn-browser-compat-data": "^0.0.96",
     "npm-registry-fetch": "^5.0.0",
     "parse5": "^5.1.0",
     "parse5-htmlparser2-tree-adapter": "^5.1.0",
     "postcss": "^7.0.18",
-    "postcss-selector-parser": "^6.0.2",
-    "postcss-value-parser": "^4.0.2",
     "request": "^2.88.0",
     "semver": "^6.3.0",
     "strip-bom": "^4.0.0",
@@ -99,7 +96,6 @@
     "lint:js": "eslint . --cache --ext js --ext md --ext ts --ignore-path ../../.eslintignore --report-unused-disable-directives",
     "lint:md": "node ../../scripts/lint-markdown.js",
     "prebuild": "npm-run-all prebuild:*",
-    "prebuild:mdn": "node ./scripts/mdn-browser-compat-data.js",
     "prebuild:mime-db": "node ./scripts/mime-db.js",
     "test": "npm run prebuild && npm run lint && npm run build && npm run test-only",
     "test-only": "nyc ava",
