--- packages/utils/src/compat/.gitignore
@@ -1,1 +1,0 @@
-browser-compat-data.ts
